make sure mocha tests can all run together
