+++ lib/utils.js
@@ -133,7 +133,7 @@ function Utils(config) {
             south: truncateNum(info.south),
             east: truncateNum(info.east),
             north: truncateNum(info.north),
-            s3Url: useS3 ? ['s3:/', config.bucket, s3Key].join('/') : null
+            s3url: useS3 ? ['s3:/', config.bucket, s3Key].join('/') : null
         };
 
         var search = {
